Send only published posts by default
Only author can interact with unpublished and archived post. To do so
they need to specify query `?author=me` to enable authorization.
If the query `?author=me` is not included in url, then authorization
will not be performed.

It's realized in this way, because this route could be accessed without
authorization (to read published posts). That is, it's possible that
authorize() will call next(err). So authorization should be optional
here.

It could be done without query parameters, using different route
like '/account/posts' or '/myposts/' with authorization (not optional)
but it would end up in more endpoints.